requirements.txt
flask>=2.3.0
flask-cors>=4.0.0

pillow>=10.0.0

transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
safetensors>=0.4.3,<0.6.0
timm>=0.9.0

huggingface-hub>=0.17.0

requests>=2.31.0
tqdm>=4.66.0
pandas>=2.0.0

duckdb>=1.4.0
git+https://github.com/deepseek-ai/Janus.git
pyinstaller>=6.0.0

# Optional: Flash Attention (requires compatible GPU)
# Uncomment to install:
# flash-attn>=2.3.0
